Refs mcvine/devel#8. Refactor: moved funcs in phonon.from_phonopy to phonon.from_phonopy.idf
